네이버 실적분석 테이블 파싱

+++ src/main/java/me/siyoon/stockfilter/application/port/in/StockFilterCommand.java
@@ -18,10 +18,10 @@ public class StockFilterCommand {
     public static class NetIncomeCommand {
 
         public final List<Period> periods;
-        public final Long threshold; // 억원
+        public final Double threshold;
 
         @JsonCreator(mode = JsonCreator.Mode.PROPERTIES)
-        public NetIncomeCommand(List<Period> periods, Long threshold) {
+        public NetIncomeCommand(List<Period> periods, Double threshold) {
             this.periods = periods;
             this.threshold = threshold;
         }
